[frontend] refactor objective to separate component
